+++ Includes/libft.h
@@ -64,6 +64,9 @@ int			ft_putstr_fd(char *s, int fd);
 int			ft_putendl_fd(char *s, int fd);
 int			ft_putnbr_fd(int n, int fd);
 
+size_t  	ft_arr_size(char **arr);
+char    	**ft_arr_copy(char **envp, size_t size);
+
 t_list		*ft_lstnew(void *content);
 void		ft_lstadd_front(t_list **lst, t_list *new);
 int			ft_lstsize(t_list *lst);
