+++ Sources/Environment/get_paths.c
