+++ Sources/Environment/get_svar.c
