+++ Sources/Execution/Built_in/Export/export_utils.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../Includes/minishell.h"
+#include "minishell.h"
 
 int	var_exist(t_mshell *shell, char *var)
 {
