+++ Sources/Execution/heredoc.c
@@ -76,55 +76,6 @@ expand*/
 // 	free(join);
 // }
 
-void	heredoc(char *delimiter, int fd_in, t_envp *envp)
-{
-	char	*line;
-	char	*result;
-	char	*temp;
-	
-	result = NULL;
-	line = NULL;
-	while (1)
-	{
-		line = readline(">");
-		if (!ft_strcmp(line, delimiter))
-		{
-			ft_dprintf(fd_in, "%s", result);
-			ft_free_null(&result);
-			ft_free_null(&line);
-			close(fd_in);
-			return ;
-		}
-		else
-		{
-			if (find_char_index(line, '$') >= 0)
-			{
-				if (check_after_dollar(line))
-				{
-    				temp = expand_envvar(line, envp);
-					ft_free_null(&line);
-				}
-			}
-			else
-				temp = ft_strdup(line);
-			line = strjoin_free_first(temp, "\n");
-			if (result)
-			{
-				result = strjoin_free_both(result, line);
-				// result = NULL;
-				line = NULL;
-			}
-			else
-			{
-				result = ft_strdup(line);
-				ft_free_null(&line);
-			}
-		}
-		if (line)
-			ft_free_null(&line);
-	}
-}
-
 // void	heredocB(char *delimiter, int fd_in, t_envp *envp)
 // {
 // 	char	*line;
