+++ Sources/Parsing/expand.c
@@ -33,28 +33,6 @@ char	*get_quotes(char *str, int c, size_t *i)
 	return (result);
 }
 
-int	are_all_quotes_closed(char *str)
-{
-	size_t	i;
-	size_t	j;
-
-	i = 0;
-	while (str[i])
-	{
-		if ((str[i] == '\'' || str[i] == '\"'))
-		{
-			j = i + 1;
-			while (str[j] && str[j] != str[i])
-				j++;
-			if (!str[j])
-				return (0);
-			i = j;
-		}
-		i++;
-	}
-	return (1);
-}
-
 int	check_after_dollar(char *str)
 {
 	size_t	i;
@@ -97,55 +75,6 @@ size_t	last_envvar_char(char *str)
 	return (i);
 }
 
-// void	add_to_list(t_list **lst, char *str)
-// {
-// 	t_list	*new;
-// 	t_list	*temp;
-
-// 	new = malloc(sizeof(t_list));
-// 	if (!(*lst))
-// 		*lst = new;
-// 	else
-// 	{
-// 		temp = *lst;
-// 		while (temp)
-// 			temp = temp->next;
-// 	}
-// 	(char *)new->content = ft_strdup(str);
-// 	temp->next = NULL;
-// 	free(str); 
-// }
-
-/*creer liste chainee
-  join les differents nodes*/
-// char	*expand_envvar(char *str, t_envp *envp)
-// {
-// 	size_t	i;
-// 	size_t	j;
-// 	t_list	*lst;
-// 	char	*res;
-// 	char	*envvar;
-
-// 	res = NULL;
-// 	i = 0;
-// 	while (str[i])
-// 	{
-// 		j = i;
-// 		if (find_char_index(&str[i], '$') >= 0)
-// 		{
-// 			i += find_char_index(&str[i], '$');
-// 			add_to_lst(&lst, ft_substr(str, j, i - j));
-// 			i++;
-// 			j = 0;
-// 			i += last_envvar_char(&str[i]);
-// 			temp = ft_substr(str, j, i - j);
-// 			if (!temp)
-// 				temp = ft_strdup("");
-			
-// 		}
-// 	}
-// }
-
 char	*expand_envvar(char *str, t_envp *envp)
 {
 	size_t	i;
@@ -174,7 +103,7 @@ char	*expand_envvar(char *str, t_envp *envp)
 			if (!temp)
 				temp = ft_strdup("");
 			join = strjoin_free_both(join, get_envvar_content(envp, temp));
-			ft_free_null(temp);
+			ft_free_null(&temp);
 			if (res)
 				res = strjoin_free_first(res, join);
 			else
@@ -193,7 +122,7 @@ char	*expand_envvar(char *str, t_envp *envp)
 		}
 	}
 	if (join)
-		ft_free_null(join);
+		ft_free_null(&join);
 	return (res);
 }
