+++ Sources/Parsing/manage_quotes.c
